+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/analytics_layout.test.tsx
@@ -8,18 +8,22 @@
 import '../../../__mocks__/shallow_useeffect.mock';
 import { mockKibanaValues, setMockValues, setMockActions } from '../../../__mocks__/kea_logic';
 import { mockUseParams } from '../../../__mocks__/react_router';
+import '../../__mocks__/engine_logic.mock';
 
 import React from 'react';
 
 import { shallow } from 'enzyme';
 
-import { FlashMessages } from '../../../shared/flash_messages';
-import { Loading } from '../../../shared/loading';
-import { rerender } from '../../../test_helpers';
-import { LogRetentionCallout } from '../log_retention';
+import {
+  rerender,
+  getPageTitle,
+  getPageHeaderActions,
+  getPageHeaderChildren,
+} from '../../../test_helpers';
+import { LogRetentionTooltip, LogRetentionCallout } from '../log_retention';
 
 import { AnalyticsLayout } from './analytics_layout';
-import { AnalyticsHeader } from './components';
+import { AnalyticsFilters } from './components';
 
 describe('AnalyticsLayout', () => {
   const { history } = mockKibanaValues;
@@ -47,18 +51,20 @@ describe('AnalyticsLayout', () => {
       </AnalyticsLayout>
     );
 
-    expect(wrapper.find(FlashMessages)).toHaveLength(1);
     expect(wrapper.find(LogRetentionCallout)).toHaveLength(1);
+    expect(getPageHeaderActions(wrapper).find(LogRetentionTooltip)).toHaveLength(1);
+    expect(getPageHeaderChildren(wrapper).find(AnalyticsFilters)).toHaveLength(1);
 
-    expect(wrapper.find(AnalyticsHeader).prop('title')).toEqual('Hello');
+    expect(getPageTitle(wrapper)).toEqual('Hello');
     expect(wrapper.find('[data-test-subj="world"]').text()).toEqual('World!');
+
+    expect(wrapper.prop('pageChrome')).toEqual(['Engines', 'some-engine', 'Analytics']);
   });
 
-  it('renders a loading component if data is not done loading', () => {
-    setMockValues({ ...values, dataLoading: true });
-    const wrapper = shallow(<AnalyticsLayout title="" />);
+  it('passes analytics breadcrumbs', () => {
+    const wrapper = shallow(<AnalyticsLayout title="Some page" breadcrumbs={['Queries']} />);
 
-    expect(wrapper.type()).toEqual(Loading);
+    expect(wrapper.prop('pageChrome')).toEqual(['Engines', 'some-engine', 'Analytics', 'Queries']);
   });
 
   describe('data loading', () => {
