+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/analytics_layout.tsx
@@ -10,25 +10,27 @@ import { useParams } from 'react-router-dom';
 
 import { useValues, useActions } from 'kea';
 
-import { EuiSpacer } from '@elastic/eui';
-
-import { FlashMessages } from '../../../shared/flash_messages';
 import { KibanaLogic } from '../../../shared/kibana';
-import { Loading } from '../../../shared/loading';
+import { BreadcrumbTrail } from '../../../shared/kibana_chrome/generate_breadcrumbs';
+import { getEngineBreadcrumbs } from '../engine';
+import { AppSearchPageTemplate } from '../layout';
 
-import { LogRetentionCallout, LogRetentionOptions } from '../log_retention';
+import { LogRetentionTooltip, LogRetentionCallout, LogRetentionOptions } from '../log_retention';
 
-import { AnalyticsHeader } from './components';
+import { AnalyticsFilters } from './components';
+import { ANALYTICS_TITLE } from './constants';
 
 import { AnalyticsLogic } from './';
 
 interface Props {
   title: string;
+  breadcrumbs?: BreadcrumbTrail;
   isQueryView?: boolean;
   isAnalyticsView?: boolean;
 }
 export const AnalyticsLayout: React.FC<Props> = ({
   title,
+  breadcrumbs = [],
   isQueryView,
   isAnalyticsView,
   children,
@@ -43,15 +45,21 @@ export const AnalyticsLayout: React.FC<Props> = ({
     if (isAnalyticsView) loadAnalyticsData();
   }, [history.location.search]);
 
-  if (dataLoading) return <Loading />;
-
   return (
-    <>
-      <AnalyticsHeader title={title} />
-      <FlashMessages />
+    <AppSearchPageTemplate
+      pageChrome={getEngineBreadcrumbs([ANALYTICS_TITLE, ...breadcrumbs])}
+      pageHeader={{
+        pageTitle: title,
+        rightSideItems: [
+          <LogRetentionTooltip type={LogRetentionOptions.Analytics} position="left" />,
+        ],
+        children: <AnalyticsFilters />,
+        responsive: false,
+      }}
+      isLoading={dataLoading}
+    >
       <LogRetentionCallout type={LogRetentionOptions.Analytics} />
       {children}
-      <EuiSpacer />
-    </>
+    </AppSearchPageTemplate>
   );
 };
