+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/analytics_router.tsx
@@ -9,7 +9,6 @@ import React from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
 
 import { APP_SEARCH_PLUGIN } from '../../../../../common/constants';
-import { SetAppSearchChrome as SetPageChrome } from '../../../shared/kibana_chrome';
 import { NotFound } from '../../../shared/not_found';
 import {
   ENGINE_ANALYTICS_PATH,
@@ -23,14 +22,7 @@ import {
 } from '../../routes';
 import { generateEnginePath, getEngineBreadcrumbs } from '../engine';
 
-import {
-  ANALYTICS_TITLE,
-  TOP_QUERIES,
-  TOP_QUERIES_NO_RESULTS,
-  TOP_QUERIES_NO_CLICKS,
-  TOP_QUERIES_WITH_CLICKS,
-  RECENT_QUERIES,
-} from './constants';
+import { ANALYTICS_TITLE } from './constants';
 import {
   Analytics,
   TopQueries,
@@ -42,42 +34,37 @@ import {
 } from './views';
 
 export const AnalyticsRouter: React.FC = () => {
-  const ANALYTICS_BREADCRUMB = getEngineBreadcrumbs([ANALYTICS_TITLE]);
-
   return (
     <Switch>
       <Route exact path={ENGINE_ANALYTICS_PATH}>
-        <SetPageChrome trail={ANALYTICS_BREADCRUMB} />
         <Analytics />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_TOP_QUERIES_PATH}>
-        <SetPageChrome trail={[...ANALYTICS_BREADCRUMB, TOP_QUERIES]} />
         <TopQueries />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_TOP_QUERIES_NO_RESULTS_PATH}>
-        <SetPageChrome trail={[...ANALYTICS_BREADCRUMB, TOP_QUERIES_NO_RESULTS]} />
         <TopQueriesNoResults />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_TOP_QUERIES_NO_CLICKS_PATH}>
-        <SetPageChrome trail={[...ANALYTICS_BREADCRUMB, TOP_QUERIES_NO_CLICKS]} />
         <TopQueriesNoClicks />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_TOP_QUERIES_WITH_CLICKS_PATH}>
-        <SetPageChrome trail={[...ANALYTICS_BREADCRUMB, TOP_QUERIES_WITH_CLICKS]} />
         <TopQueriesWithClicks />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_RECENT_QUERIES_PATH}>
-        <SetPageChrome trail={[...ANALYTICS_BREADCRUMB, RECENT_QUERIES]} />
         <RecentQueries />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_QUERY_DETAIL_PATH}>
-        <QueryDetail breadcrumbs={ANALYTICS_BREADCRUMB} />
+        <QueryDetail />
       </Route>
       <Route exact path={ENGINE_ANALYTICS_QUERY_DETAILS_PATH}>
         <Redirect to={generateEnginePath(ENGINE_ANALYTICS_PATH)} />
       </Route>
       <Route>
-        <NotFound breadcrumbs={ANALYTICS_BREADCRUMB} product={APP_SEARCH_PLUGIN} />
+        <NotFound
+          breadcrumbs={getEngineBreadcrumbs([ANALYTICS_TITLE])}
+          product={APP_SEARCH_PLUGIN}
+        />
       </Route>
     </Switch>
   );
