--- x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/components/analytics_header.scss
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-.analyticsHeader {
-  flex-wrap: wrap;
-
-  &__filters.euiPageHeaderSection {
-    width: 100%;
-    margin: $euiSizeM 0;
-  }
-}
