+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/views/recent_queries.tsx
@@ -18,7 +18,7 @@ export const RecentQueries: React.FC = () => {
   const { recentQueries } = useValues(AnalyticsLogic);
 
   return (
-    <AnalyticsLayout isAnalyticsView title={RECENT_QUERIES}>
+    <AnalyticsLayout isAnalyticsView title={RECENT_QUERIES} breadcrumbs={[RECENT_QUERIES]}>
       <AnalyticsSearch />
       <RecentQueriesTable items={recentQueries} />
     </AnalyticsLayout>
