+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/analytics/views/top_queries_no_clicks.tsx
@@ -18,7 +18,11 @@ export const TopQueriesNoClicks: React.FC = () => {
   const { topQueriesNoClicks } = useValues(AnalyticsLogic);
 
   return (
-    <AnalyticsLayout isAnalyticsView title={TOP_QUERIES_NO_CLICKS}>
+    <AnalyticsLayout
+      isAnalyticsView
+      title={TOP_QUERIES_NO_CLICKS}
+      breadcrumbs={[TOP_QUERIES_NO_CLICKS]}
+    >
       <AnalyticsSearch />
       <AnalyticsTable items={topQueriesNoClicks} hasClicks />
     </AnalyticsLayout>
